Remove controller methods for manually adding new payment, removing old/existing payment; Update accordingly

+++ backend/src/main/java/com/eventhive/payments/PaymentDTO.java
@@ -4,15 +4,15 @@ import java.time.Instant;
 import java.util.UUID;
 
 public record PaymentDTO(
-        UUID id,
-        String stripePaymentIntentId,
-        Integer amountCents,
-        String currency,
-        String status,
-        Instant purchasedAt,
-        Instant refundedAt,
-        Instant createdAt,
-        Instant updatedAt,
-        UUID bookingId) {
+                UUID id,
+                String stripePaymentIntentId,
+                Long amountCents,
+                String currency,
+                String status,
+                Instant purchasedAt,
+                Instant refundedAt,
+                Instant createdAt,
+                Instant updatedAt,
+                UUID bookingId) {
 
 }
